Store permissionName as additional key for permission lookups (#168)
This allows the pset files to specify values as either `displayName` 
(as they always have) or as `permissionName`, which allows them to
include permissions which do not have a `displayName` attribute. 
Indeed, many permissions _do not have_ a `displayName` because
the UI no longer presents `displayName` (it used to, but now uses
a translation keyed to `permissionName` in order to support l10n
of those values).