correctly handle single label
When multiple `--label value` flags are passed in, `label` is converted
to an array of strings, but when only one is passed in it remains a
plain string. The Jira API expects an array, hence, `label` must always
be converted to an array.

If it is a string, split it on commas, allowing usage like either
`--label foo --label bar` or `--label foo,bar` to have the same effect.