remove parse-matrix
The project-team-po responsibility matrix can no longer be parsed
due to being reformatted with a "Table Filter" macro that leverages
iframes in complicated ways that don't lend themselves to easy parsing.
Alas. The silver lining is that it can easily be exported to CSV, but
not at a stable URL :exploding_head: so the jspam tool has been
rewritten to include that file locally, while this tool is simply being
removed.