FOLIO-4362 handle patch values > 8, build values > 999 (#349)
The previous shell script concatenated a bunch of values together,
handling numbers as strings rather than as numbers, eventually leading
to a problem where a sufficiently large patch value (ok, it was 9) and
sufficiently large build number (>999) combined to create the version
string `11.0.9099000000001021`, and the patch value there is greater
than 2^53 - 1, which turns out to be the undocumented max as of npm
v10.9.2.

This refactor into node preserves most of the existing behavior but
manages numbers as numbers so that build numbers like 12, 123, 1234
generate patch values like 1090012, 1090123, 1091234 respectively,
and adds some boundary checking (patch values max out at 900, build
numbers at 999_999_999 or 9_999_999_999 depending on how the script is
called).

Additional details, history, and gotchas are in the comments on the
script itself.

Refs FOLIO-4362